requirements.txt
blinker==1.9.0
certifi==2025.8.3
click==8.1.8
Flask==3.1.1
flask-cors==6.0.1
gunicorn==23.0.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
sib-api-v3-sdk==7.6.0
six==1.17.0
urllib3==2.5.0
Werkzeug==3.1.3
zipp==3.23.0
